# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Golang Project Blueprint.",
        "title": "Blueprint Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "martin7.heinz@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/appointment/BookAppintment": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "creates new appointment request",
                "parameters": [
                    {
                        "description": "Appointment request",
                        "name": "guest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messageentities.BookAppointment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/messageentities.ResponseManager"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/guest/GetVisitorByEmail/{emailaddress}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "fetch visitor by email address",
                "parameters": [
                    {
                        "description": "get guest details by email request",
                        "name": "guest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messageentities.GetVisitorsByEmail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/messageentities.VisitorsDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/office/CreateOffice": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create new office",
                "parameters": [
                    {
                        "description": "office creation request",
                        "name": "office",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messageentities.CreateOffice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/messageentities.ResponseManager"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/guest/CreateVisitor": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "creates new student",
                "parameters": [
                    {
                        "description": "Create new visitor",
                        "name": "visitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messageentities.VisitorsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/messageentities.ResponseManager"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "messageentities.BookAppointment": {
            "type": "object",
            "properties": {
                "AppointmentTime": {
                    "type": "string"
                },
                "AppointmentType": {
                    "type": "string"
                },
                "Floor": {
                    "type": "string"
                },
                "Location": {
                    "type": "string"
                },
                "MeetingRoom": {
                    "type": "string"
                },
                "ProposedDate": {
                    "type": "string"
                },
                "ProposedDuration": {
                    "type": "string"
                },
                "Purpose": {
                    "type": "string"
                },
                "ScheduleType": {
                    "description": "one-off, daily",
                    "type": "string"
                },
                "StaffEmail": {
                    "type": "string"
                },
                "VisitorEmail": {
                    "type": "string"
                }
            }
        },
        "messageentities.CreateOffice": {
            "type": "object",
            "properties": {
                "Floor": {
                    "type": "string"
                },
                "OfficeAddress": {
                    "type": "string"
                },
                "OfficeDepartment": {
                    "type": "string"
                },
                "OfficeName": {
                    "type": "string"
                },
                "OfficePhoneNumber": {
                    "type": "string"
                },
                "OfficeUnit": {
                    "type": "string"
                }
            }
        },
        "messageentities.GetVisitorsByEmail": {
            "type": "object",
            "properties": {
                "VisitorEmail": {
                    "type": "string"
                }
            }
        },
        "messageentities.ResponseManager": {
            "type": "object",
            "properties": {
                "ResponseCode": {
                    "type": "string"
                },
                "ResponseDescription": {
                    "type": "string"
                }
            }
        },
        "messageentities.VisitorsDetails": {
            "type": "object",
            "properties": {
                "City": {
                    "type": "string"
                },
                "Company": {
                    "type": "string"
                },
                "EmailAddress": {
                    "type": "string"
                },
                "FullName": {
                    "type": "string"
                },
                "HomeAddress": {
                    "type": "string"
                },
                "PhoneNumber": {
                    "type": "string"
                },
                "State": {
                    "type": "string"
                }
            }
        },
        "messageentities.VisitorsRequest": {
            "type": "object",
            "properties": {
                "City": {
                    "type": "string"
                },
                "Company": {
                    "type": "string"
                },
                "EmailAddress": {
                    "type": "string"
                },
                "FullName": {
                    "type": "string"
                },
                "HomeAddress": {
                    "type": "string"
                },
                "PhoneNumber": {
                    "type": "string"
                },
                "State": {
                    "type": "string"
                }
            }
        }
    }
}